Fix some minor bugs with the linear gen and tf_model

+++ data_gen/linear_gen.py
@@ -6,7 +6,7 @@ class LinearGen(BaseGen):
         super().__init__()
 
     def create_batch(self, w, b, stddev, size=100):
-        w = tf.constant(w)
+        w = tf.cast(tf.constant(w), tf.float32)
         X = tf.zeros((size, w.shape[0]))
         X += tf.random.normal(shape=X.shape)
         y = tf.matmul(X, tf.reshape(w, (-1, 1))) + b
